+++ src/index.ts
@@ -31,9 +31,11 @@ export { Patreon } from "./providers/patreon.js";
 export { Reddit } from "./providers/reddit.js";
 export { Roblox } from "./providers/roblox.js";
 export { Salesforce } from "./providers/salesforce.js";
+export { Shikimori } from "./providers/shikimori.js";
 export { Slack } from "./providers/slack.js";
 export { Spotify } from "./providers/spotify.js";
 export { Strava } from "./providers/strava.js";
+export { Tiltify } from "./providers/tiltify.js";
 export { Tumblr } from "./providers/tumblr.js";
 export { Twitch } from "./providers/twitch.js";
 export { Twitter } from "./providers/twitter.js";
@@ -42,6 +44,7 @@ export { WorkOS } from "./providers/workos.js";
 export { Yahoo } from "./providers/yahoo.js";
 export { Yandex } from "./providers/yandex.js";
 export { Zoom } from "./providers/zoom.js";
+export { FortyTwo } from "./providers/42.js";
 
 export type {
 	AmazonCognitoRefreshedTokens,
@@ -79,9 +82,11 @@ export type { PatreonTokens } from "./providers/patreon.js";
 export type { RedditTokens } from "./providers/reddit.js";
 export type { RobloxTokens } from "./providers/roblox.js";
 export type { SalesforceTokens } from "./providers/salesforce.js";
+export type { ShikimoriTokens } from "./providers/shikimori.js";
 export type { SlackTokens } from "./providers/slack.js";
 export type { SpotifyTokens } from "./providers/spotify.js";
 export type { StravaTokens } from "./providers/strava.js";
+export type { TiltifyTokens } from "./providers/tiltify.js";
 export type { TumblrTokens } from "./providers/tumblr.js";
 export type { TwitchTokens } from "./providers/twitch.js";
 export type { TwitterTokens } from "./providers/twitter.js";
@@ -90,6 +95,7 @@ export type { WorkOSTokens } from "./providers/workos.js";
 export type { YahooTokens } from "./providers/yahoo.js";
 export type { YandexTokens } from "./providers/yandex.js";
 export type { ZoomTokens } from "./providers/zoom.js";
+export type { FortyTwoTokens } from "./providers/42.js";
 
 export { generateCodeVerifier, generateState, OAuth2RequestError } from "oslo/oauth2";