+++ src/providers/github.ts
@@ -52,3 +52,4 @@ export class GitHub implements OAuth2Provider {
 export interface GitHubTokens {
 	accessToken: string;
 }
+