+++ src/providers/linkedin.ts
@@ -31,25 +31,31 @@ export class LinkedIn implements OAuth2Provider {
 	}
 
 	public async validateAuthorizationCode(code: string): Promise<LinkedInTokens> {
-		const result = await this.client.validateAuthorizationCode<TokenResponseBody>(code, {
-			authenticateWith: "request_body",
-			credentials: this.clientSecret
-		});
+		const result = await this.client.validateAuthorizationCode<AuthorizationCodeResponseBody>(
+			code,
+			{
+				authenticateWith: "request_body",
+				credentials: this.clientSecret
+			}
+		);
 		const tokens: LinkedInTokens = {
+			idToken: result.id_token,
 			accessToken: result.access_token,
 			accessTokenExpiresAt: createDate(new TimeSpan(result.expires_in, "s")),
-			refreshToken: result.refresh_token,
-			refreshTokenExpiresAt: createDate(new TimeSpan(result.refresh_token_expires_in, "s"))
+			refreshToken: result.refresh_token ?? null,
+			refreshTokenExpiresAt: result.refresh_token_expires_in
+				? createDate(new TimeSpan(result.refresh_token_expires_in, "s"))
+				: null
 		};
 		return tokens;
 	}
 
-	public async refreshAccessToken(accessToken: string): Promise<LinkedInTokens> {
-		const result = await this.client.refreshAccessToken<TokenResponseBody>(accessToken, {
+	public async refreshAccessToken(accessToken: string): Promise<LinkedInRefreshedTokens> {
+		const result = await this.client.refreshAccessToken<RefreshTokenResponseBody>(accessToken, {
 			authenticateWith: "request_body",
 			credentials: this.clientSecret
 		});
-		const tokens: LinkedInTokens = {
+		const tokens: LinkedInRefreshedTokens = {
 			accessToken: result.access_token,
 			accessTokenExpiresAt: createDate(new TimeSpan(result.expires_in, "s")),
 			refreshToken: result.refresh_token,
@@ -59,7 +65,15 @@ export class LinkedIn implements OAuth2Provider {
 	}
 }
 
-interface TokenResponseBody {
+interface AuthorizationCodeResponseBody {
+	id_token: string;
+	access_token: string;
+	expires_in: number;
+	refresh_token?: string; // available only if your application is authorized for programmatic refresh tokens
+	refresh_token_expires_in?: number;
+}
+
+interface RefreshTokenResponseBody {
 	access_token: string;
 	expires_in: number;
 	refresh_token: string;
@@ -67,6 +81,14 @@ interface TokenResponseBody {
 }
 
 export interface LinkedInTokens {
+	idToken: string;
+	accessToken: string;
+	accessTokenExpiresAt: Date;
+	refreshToken: string | null;
+	refreshTokenExpiresAt: Date | null;
+}
+
+export interface LinkedInRefreshedTokens {
 	accessToken: string;
 	accessTokenExpiresAt: Date;
 	refreshToken: string;