+++ src/providers/okta.ts
@@ -51,9 +51,9 @@ export class Okta implements OAuth2ProviderWithPKCE {
 
 	public async validateAuthorizationCode(code: string, codeVerifier: string): Promise<OktaTokens> {
 		const result = await this.client.validateAuthorizationCode<TokenResponseBody>(code, {
-			authenticateWith: "request_body",
 			codeVerifier,
-			credentials: this.clientSecret
+			credentials: this.clientSecret,
+			authenticateWith: "request_body"
 		});
 
 		const tokens: OktaTokens = {
@@ -74,8 +74,8 @@ export class Okta implements OAuth2ProviderWithPKCE {
 		}
 	): Promise<OktaTokens> {
 		const result = await this.client.refreshAccessToken<TokenResponseBody>(refreshToken, {
-			authenticateWith: "request_body",
 			credentials: this.clientSecret,
+			authenticateWith: "request_body",
 			scopes: options?.scopes ?? []
 		});